feat: rename Tolgee file to TolgeeCore
