+++ packages/core/src/TolgeeCore.ts
