+++ packages/core/src/index.ts
@@ -1,5 +1,5 @@
 export { getFallback, getFallbackArray } from './helpers';
-export { TolgeeCore } from './Tolgee';
+export { TolgeeCore } from './TolgeeCore';
 export * from './types';
 export { getTranslateProps } from './TranslateParams';
 export { FormatSimple } from './FormatSimple/FormatSimple';
